+++ lib/src/ast/css/media_query.dart
@@ -2,6 +2,7 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import '../../logger.dart';
 import '../../parse/media_query.dart';
 import '../../utils.dart';
 
@@ -28,8 +29,8 @@ class CssMediaQuery {
   /// If passed, [url] is the name of the file from which [contents] comes.
   ///
   /// Throws a [SassFormatException] if parsing fails.
-  static List<CssMediaQuery> parseList(String contents, {url}) =>
-      new MediaQueryParser(contents, url: url).parse();
+  static List<CssMediaQuery> parseList(String contents, {url, Logger logger}) =>
+      new MediaQueryParser(contents, url: url, logger: logger).parse();
 
   /// Creates a media query specifies a type and, optionally, features.
   CssMediaQuery(this.type, {this.modifier, Iterable<String> features})