+++ lib/src/ast/selector/simple.dart
@@ -3,6 +3,7 @@
 // https://opensource.org/licenses/MIT.
 
 import '../../exception.dart';
+import '../../logger.dart';
 import '../../parse/selector.dart';
 import '../selector.dart';
 
@@ -34,8 +35,9 @@ abstract class SimpleSelector extends Selector {
   ///
   /// Throws a [SassFormatException] if parsing fails.
   factory SimpleSelector.parse(String contents,
-          {url, bool allowParent: true}) =>
-      new SelectorParser(contents, url: url, allowParent: allowParent)
+          {url, Logger logger, bool allowParent: true}) =>
+      new SelectorParser(contents,
+              url: url, logger: logger, allowParent: allowParent)
           .parseSimpleSelector();
 
   /// Returns a new [SimpleSelector] based on [this], as though it had been