+++ lib/src/parse/at_root_query.dart
@@ -5,11 +5,13 @@
 import 'package:charcode/charcode.dart';
 
 import '../ast/sass.dart';
+import '../logger.dart';
 import 'parser.dart';
 
 /// A parser for `@at-root` queries.
 class AtRootQueryParser extends Parser {
-  AtRootQueryParser(String contents, {url}) : super(contents, url: url);
+  AtRootQueryParser(String contents, {url, Logger logger})
+      : super(contents, url: url, logger: logger);
 
   AtRootQuery parse() {
     return wrapSpanFormatException(() {